+++ fsm/state_machine.gd
@@ -21,22 +21,51 @@ extends Node
 # -- SIGNALS ------------------------------------------------------------------------- #
 
 ## Emitted when a 'State' is entered.
-signal state_entered(next)
+signal state_entered(path: NodePath)
 
 ## Emitted when a 'State' is exited.
-signal state_exited(previous)
+signal state_exited(path: NodePath)
+
+## Emitted when a 'State' transition is started.
+signal transition_started(from: NodePath, to: NodePath)
+
+## Emitted when a 'State' transition is finished.
+signal transition_finished(from: NodePath, to: NodePath)
+
+# -- DEFINITIONS --------------------------------------------------------------------- #
+
+enum StateMachineProcessCallback {
+	STATE_MACHINE_PROCESS_PHYSICS = 0, STATE_MACHINE_PROCESS_IDLE = 1
+}
 
 # -- DEPENDENCIES -------------------------------------------------------------------- #
 
 const Iterators := preload("../iter/node.gd")
 const State := preload("state.gd")
 
-# -- DEFINITIONS --------------------------------------------------------------------- #
-
 # -- CONFIGURATION ------------------------------------------------------------------- #
 
 ## The starting 'State' for the 'StateMachine'; will be transitioned to on 'ready'.
 @export_node_path var initial: NodePath
+
+## Whether to "compact" the 'StateMachine' by extracting 'State' scripts as 'Object'
+## instances from each child 'Node'.
+@export var compact: bool = true
+
+## process_callback determines whether 'update' is called during the physics or idle
+## process callback function (if the process mode allows for it).
+@export
+var process_callback := StateMachineProcessCallback.STATE_MACHINE_PROCESS_PHYSICS:
+	set(value):
+		process_callback = value
+
+		match value:
+			StateMachineProcessCallback.STATE_MACHINE_PROCESS_PHYSICS:
+				set_physics_process(true)
+				set_process(false)
+			StateMachineProcessCallback.STATE_MACHINE_PROCESS_IDLE:
+				set_physics_process(false)
+				set_process(true)
 
 # -- INITIALIZATION ------------------------------------------------------------------ #
 
@@ -95,13 +124,23 @@ func update(delta: float) -> void:
 
 
 func _enter_tree() -> void:
-	assert(initial, "Invalid configuration; missing 'initial' property!")
+	assert(initial, "invalid configuration; missing 'initial' property")
 
 	# Iterate through all 'State' nodes
 	for n in Iterators.descendents(
 		self, Iterators.Filter.ALL, Iterators.Order.DEPTH_FIRST
 	):
-		var s := _extract_state(n)
+		var s: State
+
+		if compact:
+			s = _extract_state(n)
+		else:
+			s = (n as Object) as State
+
+		if s == null:
+			continue
+
+		assert(s != null, "child node is not a valid state")
 
 		var p := n.get_parent()
 		s._parent = _states[get_path_to(p)] if p and p != self else null
@@ -111,36 +150,49 @@ func _enter_tree() -> void:
 		_leaves[s.get_instance_id()] = s
 		_states[get_path_to(n)] = s
 
-	# Delete 'Node' instances to prevent their addition to the scene.
-	var index := 0
-	while index < get_child_count():
-		var child := get_child(index)
-		remove_child(child)
-		child.free()
+		print(
+			"std/fsm/state_machine.gd[",
+			get_instance_id(),
+			"]: registered state: " + str(s._path)
+		)
 
-	# Transition to the initial 'State'
-	_transition_to(initial)
-	assert(state is State, "Failed to set initial 'State'!")
-	assert(
-		_leaves.has(state.get_instance_id()),
-		"Invalid configuration; 'initial' is not a leaf 'State'!"
-	)
+	# Delete 'Node' instances to prevent their addition to the scene.
+	if compact:
+		var index := 0
+		while index < get_child_count():
+			var child := get_child(index)
+			remove_child(child)
+			child.free()
 
 
 func _notification(what) -> void:
-	if what == NOTIFICATION_PREDELETE:
+	if compact and what == NOTIFICATION_PREDELETE:
 		state = null
 		for s in _states.values():
 			if is_instance_valid(s):
 				s.free()
 
 
-# NOTE: To disable auto-'update' calls, set 'process_mode' to 'PROCESS_MODE_DISABLED'.
 func _physics_process(delta) -> void:
 	update(delta)
 
 
-# -- PRIVATE METHODS (OVERRIDES) ----------------------------------------------------- #
+func _process(delta) -> void:
+	update(delta)
+
+
+func _ready() -> void:
+	# Trigger the setter to properly configure callback functions.
+	process_callback = process_callback
+
+	# Transition to the initial 'State'
+	_transition_to(initial)
+	assert(state is State, "failed to set initial 'State'")
+	assert(
+		_leaves.has(state.get_instance_id()),
+		"invalid configuration; 'initial' is not a leaf 'State'"
+	)
+
 
 # -- PRIVATE METHODS ----------------------------------------------------------------- #
 
@@ -156,7 +208,7 @@ func _extract_state(node: Node, strict: bool = true) -> State:
 		s = s.get_base_script()
 
 	if not s is Script:
-		assert(not strict, "Failed to extract 'State' from 'Node'!")
+		assert(not strict, "failed to extract 'State' from 'Node'")
 		return null
 
 	var out: State = node.get_script().new()
@@ -176,11 +228,31 @@ func _extract_state(node: Node, strict: bool = true) -> State:
 ## @args:
 ## 	path [NodePath] - A 'NodePath' (relative to 'StateMachine') to the target 'State'.
 func _transition_to(path: NodePath) -> void:
+	assert(path != NodePath(), "missing argument: path")
+
+	# If possible, normalize the provided path.
+	if not compact:
+		var target = get_node_or_null(path)
+		assert(target is State, "invalid argument; 'path' is not a State node")
+
+		path = get_path_to(target as Object as Node)
+
 	assert(path in _states, "Invalid argument; 'path' not found in 'StateMachine'!")
 	assert(not _is_in_transition, "Invalid config; nested transitions prohibited!")
 
+	print(
+		"std/fsm/state_machine.gd[",
+		get_instance_id(),
+		"]: transition to state: " + str(path)
+	)
+
 	var next: State = _states[path]
 	assert(next.get_instance_id() in _leaves, "Argument 'next' was not a leaf 'State'!")
+
+	var from_path := state._path if state else NodePath()
+	var to_path := next._path
+
+	transition_started.emit(from_path, to_path)
 
 	var to_exit := [state] if state else []
 	var to_enter := [next]
@@ -223,6 +295,4 @@ func _transition_to(path: NodePath) -> void:
 
 	_is_in_transition = false
 
-# -- SIGNAL HANDLERS ----------------------------------------------------------------- #
-
-# -- SETTERS/GETTERS ----------------------------------------------------------------- #
+	transition_finished.emit(from_path, to_path)